requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
alibabacloud-credentials==1.0.3
alibabacloud-credentials-api==1.0.0
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.1
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_util==0.3.13
aliyun-python-sdk-core==2.16.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
APScheduler==3.11.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
darabonba-core==1.0.4
dashscope==1.25.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-speech==2.34.0
googleapis-common-protos==1.72.0
greenlet==3.2.2
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hydra-configs-pytorch-lightning==0.1.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiwer==4.0.0
jmespath==0.10.0
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
librosa==0.11.0
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.2
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai-whisper==20250625
optuna==4.3.0
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-json-logger==3.3.0
python-Levenshtein==0.27.1
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
sox==1.5.0
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.14.0
tencentcloud-sdk-python==3.0.1489
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.9.0
torchaudio==2.9.0
torchcodec==0.8.1
torchelastic==0.2.2
torchmetrics==1.7.1
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.5.0
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.14
yarl==1.20.0
